Making the digests more object oriented by putting in own class.  Gitignore some node/react stuff that made it into my project from testing.
