+++ .gitignore
@@ -17,6 +17,10 @@ test-reports/
 stash/script/counter-uploader/reports/*.json
 stash/script/counter-uploader/tmp/*.json
 
+# JS library files from node/react
+node_modules
+public/packs
+
 # ######################################################
 
 #### joe made this: http://goel.io/joe
